+++ packages/contracts/contracts/test/invariant/InvariantsBase.sol
@@ -6,6 +6,8 @@ import {console} from "forge-std/console.sol";
 import {StdInvariant} from "forge-std/StdInvariant.sol";
 
 import {DepositManagerHandler} from "./actors/DepositManagerHandler.sol";
+import {WalletHandler} from "./actors/WalletHandler.sol";
+import {TokenSwapper, SwapRequest} from "../utils/TokenSwapper.sol";
 import {TestJoinSplitVerifier} from "../harnesses/TestJoinSplitVerifier.sol";
 import {TestSubtreeUpdateVerifier} from "../harnesses/TestSubtreeUpdateVerifier.sol";
 import "../utils/NocturneUtils.sol";
@@ -28,18 +30,37 @@ contract InvariantsBase is Test {
     address SCREENER_ADDRESS = vm.addr(SCREENER_PRIVKEY);
 
     DepositManagerHandler public depositManagerHandler;
+    WalletHandler public walletHandler;
 
     Wallet public wallet;
     Handler public handler;
     TestDepositManager public depositManager;
-    WETH9 public weth;
 
-    SimpleERC20Token public erc20;
-    SimpleERC721Token public erc721;
-    SimpleERC1155Token public erc1155;
+    TokenSwapper public swapper;
+
+    WETH9 public weth;
+    SimpleERC20Token public depositErc20;
+    SimpleERC721Token public depositErc721;
+    SimpleERC1155Token public depositErc1155;
+    SimpleERC20Token public swapErc20;
+    SimpleERC721Token public swapErc721;
+    SimpleERC1155Token public swapErc1155;
 
     function print_callSummary() internal view {
         depositManagerHandler.callSummary();
+        walletHandler.callSummary();
+    }
+
+    // _______________PROTOCOL_WIDE_______________
+
+    function assert_protocol_walletBalanceEqualsCompletedDepositSumMinusTransferedOutErc20()
+        internal
+    {
+        assertEq(
+            depositManagerHandler.erc20().balanceOf(address(wallet)),
+            depositManagerHandler.ghost_completeDepositSumErc20() -
+                walletHandler.ghost_totalTransferredOutOfWallet()
+        );
     }
 
     // _______________DEPOSIT_ETH_______________
@@ -63,6 +84,15 @@ contract InvariantsBase is Test {
         );
     }
 
+    function assert_deposit_walletBalanceEqualsCompletedDepositSumETH()
+        internal
+    {
+        assertEq(
+            weth.balanceOf(address(wallet)),
+            depositManagerHandler.ghost_completeDepositSumETH()
+        );
+    }
+
     function assert_deposit_allActorsBalanceSumETHEqualsRetrieveDepositSumETH()
         internal
     {
@@ -74,15 +104,6 @@ contract InvariantsBase is Test {
         }
 
         assertEq(sum, depositManagerHandler.ghost_retrieveDepositSumETH());
-    }
-
-    function assert_deposit_walletBalanceEqualsCompletedDepositSumETH()
-        internal
-    {
-        assertEq(
-            weth.balanceOf(address(wallet)),
-            depositManagerHandler.ghost_completeDepositSumETH()
-        );
     }
 
     function assert_deposit_actorBalanceAlwaysEqualsRetrievedETH() internal {
@@ -147,15 +168,6 @@ contract InvariantsBase is Test {
         assertEq(sum, depositManagerHandler.ghost_retrieveDepositSumErc20());
     }
 
-    function assert_deposit_walletBalanceEqualsCompletedDepositSumErc20()
-        internal
-    {
-        assertEq(
-            depositManagerHandler.erc20().balanceOf(address(wallet)),
-            depositManagerHandler.ghost_completeDepositSumErc20()
-        );
-    }
-
     function assert_deposit_actorBalanceAlwaysEqualsRetrievedErc20() internal {
         address[] memory allActors = depositManagerHandler.ghost_AllActors();
 
@@ -180,6 +192,37 @@ contract InvariantsBase is Test {
                 depositManagerHandler.ghost_instantiateDepositSumErc20For(
                     allActors[i]
                 )
+            );
+        }
+    }
+
+    // _______________OPERATIONS_______________
+
+    function assert_operation_totalSwapErc20ReceivedMatchesWalletBalance()
+        internal
+    {
+        assertEq(
+            swapErc20.balanceOf(address(wallet)),
+            walletHandler.ghost_totalSwapErc20Received()
+        );
+    }
+
+    function assert_operation_walletOwnsAllSwapErc721s() internal {
+        uint256[] memory ids = walletHandler.ghost_swapErc721IdsReceived();
+        for (uint256 i = 0; i < ids.length; i++) {
+            assertEq(address(wallet), swapErc721.ownerOf(ids[i]));
+        }
+    }
+
+    function assert_operation_totalSwapErc1155ReceivedMatchesWalletBalance()
+        internal
+    {
+        uint256[] memory ids = walletHandler.ghost_swapErc1155IdsReceived();
+        for (uint256 i = 0; i < ids.length; i++) {
+            uint256 id = ids[i];
+            assertEq(
+                swapErc1155.balanceOf(address(wallet), id),
+                walletHandler.ghost_totalSwapErc1155ReceivedForId(id)
             );
         }
     }