+++ packages/contracts/contracts/test/utils/NocturneUtils.sol
@@ -18,8 +18,7 @@ struct FormatOperationArgs {
     SimpleERC20Token joinSplitToken;
     SimpleERC20Token gasToken;
     uint256 root;
-    uint256 publicSpendPerJoinSplit;
-    uint256 numJoinSplits;
+    uint256[] joinSplitPublicSpends;
     EncodedAsset[] encodedRefundAssets;
     uint256 executionGasLimit;
     uint256 maxNumRefunds;
@@ -27,6 +26,12 @@ struct FormatOperationArgs {
     Action[] actions;
     bool atomicActions;
     OperationFailureType operationFailureType;
+}
+
+struct TransferRequest {
+    SimpleERC20Token token;
+    address recipient;
+    uint256 amount;
 }
 
 library NocturneUtils {
@@ -51,6 +56,17 @@ library NocturneUtils {
         for (uint256 i = 0; i < 8; i++) {
             _values[i] = uint256(4757829);
         }
+    }
+
+    function fillJoinSplitPublicSpends(
+        uint256 perJoinSplitPublicSpend,
+        uint256 numJoinSplits
+    ) internal pure returns (uint256[] memory) {
+        uint256[] memory joinSplitPublicSpends = new uint256[](numJoinSplits);
+        for (uint256 i = 0; i < numJoinSplits; i++) {
+            joinSplitPublicSpends[i] = perJoinSplitPublicSpend;
+        }
+        return joinSplitPublicSpends;
     }
 
     function formatDepositRequest(
@@ -85,22 +101,26 @@ library NocturneUtils {
         uint256 amount
     ) public pure returns (Action[] memory) {
         Action[] memory actions = new Action[](1);
-        actions[0] = formatTransferAction(token, recipient, amount);
+        actions[0] = formatTransferAction(
+            TransferRequest({
+                token: token,
+                recipient: recipient,
+                amount: amount
+            })
+        );
         return actions;
     }
 
     function formatTransferAction(
-        SimpleERC20Token token,
-        address recipient,
-        uint256 amount
+        TransferRequest memory transferRequest
     ) public pure returns (Action memory) {
         return
             Action({
-                contractAddress: address(token),
+                contractAddress: address(transferRequest.token),
                 encodedFunction: abi.encodeWithSelector(
-                    token.transfer.selector,
-                    recipient,
-                    amount
+                    transferRequest.token.transfer.selector,
+                    transferRequest.recipient,
+                    transferRequest.amount
                 )
             });
     }
@@ -115,7 +135,7 @@ library NocturneUtils {
             operationFailure == OperationFailureType.JOINSPLIT_NF_ALREADY_IN_SET
         ) {
             require(
-                args.numJoinSplits >= 2,
+                args.joinSplitPublicSpends.length >= 2,
                 "Must specify at least 2 joinsplits for JOINSPLIT_NF_ALREADY_IN_SET failure type"
             );
         }
@@ -151,8 +171,10 @@ library NocturneUtils {
         );
 
         // Setup joinsplits depending on failure type
-        JoinSplit[] memory joinSplits = new JoinSplit[](args.numJoinSplits);
-        for (uint256 i = 0; i < args.numJoinSplits; i++) {
+        JoinSplit[] memory joinSplits = new JoinSplit[](
+            args.joinSplitPublicSpends.length
+        );
+        for (uint256 i = 0; i < args.joinSplitPublicSpends.length; i++) {
             uint256 nullifierA = 0;
             uint256 nullifierB = 0;
             if (operationFailure == OperationFailureType.JOINSPLIT_NFS_SAME) {
@@ -161,14 +183,14 @@ library NocturneUtils {
             } else if (
                 operationFailure ==
                 OperationFailureType.JOINSPLIT_NF_ALREADY_IN_SET &&
-                i + 2 == args.numJoinSplits
+                i + 2 == args.joinSplitPublicSpends.length
             ) {
                 nullifierA = uint256(2 * 0x1234); // Matches last NF B
                 nullifierB = uint256(2 * i + 1);
             } else if (
                 operationFailure ==
                 OperationFailureType.JOINSPLIT_NF_ALREADY_IN_SET &&
-                i + 1 == args.numJoinSplits
+                i + 1 == args.joinSplitPublicSpends.length
             ) {
                 nullifierA = uint256(2 * i);
                 nullifierB = uint256(2 * 0x1234); // Matches 2nd to last NF A
@@ -186,7 +208,7 @@ library NocturneUtils {
                 newNoteBEncrypted: newNoteBEncrypted,
                 proof: dummyProof(),
                 encodedAsset: encodedAsset,
-                publicSpend: args.publicSpendPerJoinSplit,
+                publicSpend: args.joinSplitPublicSpends[i],
                 encSenderCanonAddrC1X: 0,
                 encSenderCanonAddrC2X: 0
             });